# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "VASARI API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/schedeA": {
      "get": {
        "tags": [
          "Schede A"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Schede A"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rs": {
                  "example": "any"
                },
                "cm": {
                  "example": "any"
                },
                "au": {
                  "example": "any"
                },
                "rv": {
                  "example": "any"
                },
                "lc": {
                  "example": "any"
                },
                "og": {
                  "example": "any"
                },
                "cd": {
                  "example": "any"
                },
                "li": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/schedeA/ogtn/{ogtn}": {
      "get": {
        "tags": [
          "Schede A"
        ],
        "description": "Cerca nelle schede tramite campo ogtn",
        "parameters": [
          {
            "name": "ogtn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/schedeA/{id}": {
      "delete": {
        "tags": [
          "Schede A"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Schede A"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Schede A"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rs": {
                  "example": "any"
                },
                "cm": {
                  "example": "any"
                },
                "au": {
                  "example": "any"
                },
                "rv": {
                  "example": "any"
                },
                "lc": {
                  "example": "any"
                },
                "og": {
                  "example": "any"
                },
                "cd": {
                  "example": "any"
                },
                "li": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/schedeOA": {
      "get": {
        "tags": [
          "Schede OA"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Schede OA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "mt": {
                  "example": "any"
                },
                "rs": {
                  "example": "any"
                },
                "co": {
                  "example": "any"
                },
                "dt": {
                  "example": "any"
                },
                "cm": {
                  "example": "any"
                },
                "au": {
                  "example": "any"
                },
                "rv": {
                  "example": "any"
                },
                "lc": {
                  "example": "any"
                },
                "og": {
                  "example": "any"
                },
                "cd": {
                  "example": "any"
                },
                "ub": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/schedeOA/{id}": {
      "get": {
        "tags": [
          "Schede OA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Schede OA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Schede OA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "mt": {
                  "example": "any"
                },
                "rs": {
                  "example": "any"
                },
                "co": {
                  "example": "any"
                },
                "dt": {
                  "example": "any"
                },
                "cm": {
                  "example": "any"
                },
                "au": {
                  "example": "any"
                },
                "rv": {
                  "example": "any"
                },
                "lc": {
                  "example": "any"
                },
                "og": {
                  "example": "any"
                },
                "cd": {
                  "example": "any"
                },
                "ub": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/schedeOA/ogtn/{ogtn}": {
      "get": {
        "tags": [
          "Schede OA"
        ],
        "description": "Ricerca tramite campo",
        "parameters": [
          {
            "name": "ogtn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/schedeGpm": {
      "post": {
        "tags": [
          "Schede Gpm"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "abstract": {
                  "example": "any"
                },
                "pi": {
                  "example": "any"
                },
                "cb": {
                  "example": "any"
                },
                "geolocation": {
                  "example": "any"
                },
                "access_group": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "Schede Gpm"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/schedeGpm/abstract/{abstract}": {
      "get": {
        "tags": [
          "Schede Gpm"
        ],
        "description": "",
        "parameters": [
          {
            "name": "abstract",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/schedeGpm/{id}": {
      "get": {
        "tags": [
          "Schede Gpm"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Schede Gpm"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pi": {
                  "example": "any"
                },
                "cb": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Schede Gpm"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}